bloggers/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
mysqlclient
gunicorn